added NavigatorBase with applyCommand implementation that dispatches passed command to corresponding method
